refactor(packaging): package library as owid.walden
